[#164829364] Adds tests and authentication to Create, Delete, Update endpoints

+++ authors/apps/articles/views.py
@@ -1,12 +1,17 @@
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
 from django.shortcuts import get_object_or_404
 
 from .models import Article
 from .serializers import ArticleSerializer
+from .permissions import ReadOnly
+from authors.apps.authentication.models import User
 
 
 class ArticleView(APIView):
+    permission_classes = (IsAuthenticated | ReadOnly,)
+
     def get(self, request):
         articles = Article.objects.all()
         serializer = ArticleSerializer(articles, many=True)
@@ -18,11 +23,18 @@ class ArticleView(APIView):
         # Create an article from the above data
         serializer = ArticleSerializer(data=article)
         if serializer.is_valid(raise_exception=True):
-            article_saved = serializer.save()
-        return Response({"success": "Article '{}' created successfully".format(article_saved.title)})
+            article_saved = serializer.save(author=self.request.user)
+        return Response({"success": "Article '{}' created successfully".format(article_saved.title)}, status=201)
 
 
 class RetrieveArticleView(APIView):
+    permission_classes = (IsAuthenticated | ReadOnly,)
+
+    def is_owner(self, current_user_id, article_author_id):
+        if article_author_id == current_user_id:
+            return True
+        return False
+
     def get(self, request, slug):
         try:
             article = Article.objects.get(slug=slug)
@@ -33,15 +45,33 @@ class RetrieveArticleView(APIView):
                 {"message": "The article requested does not exist"}, status=404)
 
     def put(self, request, slug):
-        saved_article = get_object_or_404(Article.objects.all(), slug=slug)
+        try:
+            saved_article = Article.objects.get(slug=slug)
+        except Article.DoesNotExist:
+            return Response(
+                {"message": "The article requested does not exist"}, status=404)
+
         data = request.data.get('article')
         serializer = ArticleSerializer(
             instance=saved_article, data=data, partial=True)
         if serializer.is_valid(raise_exception=True):
-            article_saved = serializer.save()
-        return Response({"success": "Article '{}' updated successfully".format(article_saved.title)})
+            if self.is_owner(saved_article.author.id, request.user.id) is True:
+                article_saved = serializer.save()
+                return Response({"success": "Article '{}' updated successfully".format(article_saved.title)})
+            response = {"message": "Only the owner can edit this article."}
+            return Response(response, status=401)
 
     def delete(self, request, slug):
-        article = get_object_or_404(Article.objects.all(), slug=slug)
-        article.delete()
-        return Response({"message": "Article `{}` has been deleted.".format(slug)}, status=204)
+        try:
+            article = Article.objects.get(slug=slug)
+        except Article.DoesNotExist:
+            return Response(
+                {"message": "The article requested does not exist"}, status=404)
+
+        if self.is_owner(article.author.id, request.user.id) is True:
+            article.delete()
+            return Response({"message": "Article `{}` has been deleted.".format(slug)}, status=200)
+
+        response = {"message": "Only the owner can delete this article."}
+        return Response(response, status=401)
+
